Cleaned up settings navigation
