--- src/routes/Dashboard/Settings.js
@@ -1,10 +1,0 @@
-import React from "react";
-import ListPopulationComponent from "../../components/SettingsPage/ListPopulationComponent";
-
-export default function Settings(props) {
-  return (
-    <div className="dashboard-children settings">
-      <ListPopulationComponent />
-    </div>
-  );
-}
